Refactored zu_verst_eink.
