+++ gettsim/taxes/eink_st.py
@@ -31,12 +31,12 @@ def _st_kein_kind_freib_tu(_st_kein_kind_freib, tu_id):
     return out.rename("_st_kein_kind_freib_tu")
 
 
-def _st_kind_freib(_zu_versteuerndes_eink_kind_freib, eink_st_params):
+def _st_kind_freib(_zu_verst_eink_kinderfreib, eink_st_params):
     """
 
     Parameters
     ----------
-    _zu_versteuerndes_eink_kind_freib
+    _zu_verst_eink_kinderfreib
     eink_st_params
 
     Returns
@@ -45,9 +45,9 @@ def _st_kind_freib(_zu_versteuerndes_eink_kind_freib, eink_st_params):
     """
     if eink_st_params["jahr"] < 2002:
         raise ValueError("Income Tax Pre 2002 not yet modelled!")
-    out = st_tarif(_zu_versteuerndes_eink_kind_freib, eink_st_params)
+    out = st_tarif(_zu_verst_eink_kinderfreib, eink_st_params)
     return pd.Series(
-        index=_zu_versteuerndes_eink_kind_freib.index, data=out, name="_st_kind_freib"
+        index=_zu_verst_eink_kinderfreib.index, data=out, name="_st_kind_freib"
     )
