+++ gettsim/tests/test_data/test_dfs_tax_sched.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","kind","_zu_verst_eink_kein_kinderfreib","_zu_versteuerndes_eink_kind_freib","_zu_versteuerndes_eink_abgelt_st_m_kein_kind_freib","_zu_versteuerndes_eink_abgelt_st_m_kind_freib","brutto_eink_5","brutto_eink_5_tu","gem_veranlagt","jahr",,"_st_kein_kind_freib_tu","_st_kind_freib_tu","_st_abgelt_st_m_kein_kind_freib","_st_abgelt_st_m_kind_freib_tu","abgelt_st_m","abgelt_st_m_tu","soli_st_m_tu","soli_st_m","G","M","S","R",,"solibase_inctax"
+"hh_id","tu_id","p_id","kind","_zu_verst_eink_kein_kinderfreib","_zu_verst_eink_kinderfreib","_zu_versteuerndes_eink_abgelt_st_m_kein_kind_freib","_zu_versteuerndes_eink_abgelt_st_m_kind_freib","brutto_eink_5","brutto_eink_5_tu","gem_veranlagt","jahr",,"_st_kein_kind_freib_tu","_st_kind_freib_tu","_st_abgelt_st_m_kein_kind_freib","_st_abgelt_st_m_kind_freib_tu","abgelt_st_m","abgelt_st_m_tu","soli_st_m_tu","soli_st_m","G","M","S","R",,"solibase_inctax"
 1,1,1,FALSE,5000,5000,5500,5500,500,500,FALSE,2018,,0,0,0,0,0,0,0.00,0.00,9000,13996,54949,260532,,0
 2,2,2,FALSE,5000,5000,6000,6000,1300,1800,TRUE,2015,,0,0,0,0,24.75,49.5,0.23,0.11,8472,13469,52881,250730,,0
 2,2,32,FALSE,5000,5000,6000,6000,500,1800,TRUE,2015,,0,0,0,0,24.75,49.5,0.23,0.11,8472,13469,52881,250730,,0
