+++ gettsim/tests/test_data/test_dfs_zve.csv
@@ -1,4 +1,4 @@
-"hh_id","tu_id","p_id","bruttolohn_m","betreuungskost_m","eink_selbstst_m","kapital_eink_m","vermiet_eink_m","jahr_renteneintr","arbeitsstunden_w","in_ausbildung","ges_rente_m","wohnort_ost","gem_veranlagt","kind","behinderungsgrad","rentenv_beit_m","prv_rente_beit_m","arbeitsl_v_beit_m","ges_krankenv_beitr_m","pflegev_beit_m","alleinerziehend","alter","anz_kinder_tu","jahr","_zu_verst_eink_kein_kinderfreib","_zu_versteuerndes_eink_kind_freib","_zu_versteuerndes_eink_abgelt_st_kein_kind_freib","_zu_versteuerndes_eink_abgelt_st_kind_freib",,"vorsorg","sum_brutto_eink","kinderfreib","sonstigevors","zve_nokfb_ind","d_zve","d_zu_kfb","zve_temp","altersfreib"
+"hh_id","tu_id","p_id","bruttolohn_m","betreuungskost_m","eink_selbstst_m","kapital_eink_m","vermiet_eink_m","jahr_renteneintr","arbeitsstunden_w","in_ausbildung","ges_rente_m","wohnort_ost","gem_veranlagt","kind","behinderungsgrad","rentenv_beit_m","prv_rente_beit_m","arbeitsl_v_beit_m","ges_krankenv_beitr_m","pflegev_beit_m","alleinerziehend","alter","anz_kinder_tu","jahr","_zu_verst_eink_kein_kinderfreib","_zu_verst_eink_kinderfreib","_zu_versteuerndes_eink_abgelt_st_kein_kind_freib","_zu_versteuerndes_eink_abgelt_st_kind_freib",,"vorsorg","sum_brutto_eink","kinderfreib","sonstigevors","zve_nokfb_ind","d_zve","d_zu_kfb","zve_temp","altersfreib"
 1,1,1,300,0,0,100,0,2043,40,"FALSE",0,"FALSE","FALSE","FALSE",0,0,0,0,0,0,"FALSE",40,0,2018,0,0,363,363,,0,0,0,0,,,,,0
 2,2,2,600,0,0,100,0,2043,40,"FALSE",0,"FALSE","FALSE","FALSE",0,60,0,12,48,9,"FALSE",40,0,2018,4842,4842,5241,5241,,1322,6200,0,804,518,,,,0
 3,3,3,900,0,0,100,0,2043,40,"FALSE",0,"FALSE","FALSE","FALSE",0,90,0,18,72,13.5,"FALSE",40,0,2018,7780,7780,8179,8179,,1984,9800,0,1207,,,,,0
